Refactor: Implement exposure reporting changes
This commit implements the changes requested in the prompt.
