+++ src/hooks/useExposureData.ts
@@ -1,0 +1,631 @@
+import { useState, useMemo } from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { supabase } from '@/integrations/supabase/client';
+import { validateAndParsePricingFormula } from '@/utils/formulaUtils';
+import { getNextMonths, formatMonthCode } from '@/utils/dateUtils';
+import { mapProductToCanonical, parsePaperInstrument } from '@/utils/productMapping';
+import { calculateNetExposure } from '@/utils/tradeUtils';
+import { 
+  MonthlyExposure,
+  ExposureData, 
+  PricingInstrument
+} from '@/types/exposure';
+
+export const useExposureData = () => {
+  const [periods] = useState<string[]>(getNextMonths(13));
+  const [visibleCategories, setVisibleCategories] = useState<string[]>(['Physical', 'Pricing', 'Paper', 'Exposure']);
+  const [selectedProducts, setSelectedProducts] = useState<string[]>([]);
+
+  // Fetch pricing instruments
+  const {
+    data: pricingInstruments = [],
+    isLoading: instrumentsLoading
+  } = useQuery({
+    queryKey: ['pricing-instruments'],
+    queryFn: async () => {
+      const {
+        data,
+        error
+      } = await supabase.from('pricing_instruments').select('id, display_name, instrument_code, is_active').eq('is_active', true);
+      if (error) throw error;
+      return data || [];
+    }
+  });
+
+  // Calculate allowed products based on pricing instruments
+  const ALLOWED_PRODUCTS = useMemo(() => {
+    const instrumentProducts = pricingInstruments.map((inst: PricingInstrument) => mapProductToCanonical(inst.display_name));
+    const biodieselProducts = ['Argus UCOME', 'Argus FAME0', 'Argus RME', 'Argus HVO'];
+    return Array.from(new Set([...instrumentProducts, ...biodieselProducts]));
+  }, [pricingInstruments]);
+
+  // Calculate biodiesel and pricing instrument products
+  const BIODIESEL_PRODUCTS = useMemo(() => {
+    return ALLOWED_PRODUCTS.filter(p => p.includes('Argus'));
+  }, [ALLOWED_PRODUCTS]);
+
+  const PRICING_INSTRUMENT_PRODUCTS = useMemo(() => {
+    return ALLOWED_PRODUCTS.filter(p => !p.includes('Argus'));
+  }, [ALLOWED_PRODUCTS]);
+
+  // Fetch trade data
+  const {
+    data: tradeData,
+    isLoading,
+    error,
+    refetch
+  } = useQuery({
+    queryKey: ['exposure-data'],
+    queryFn: async () => {
+      const {
+        data: physicalTradeLegs,
+        error: physicalError
+      } = await supabase.from('trade_legs').select(`
+          id,
+          leg_reference,
+          buy_sell,
+          product,
+          quantity,
+          pricing_formula,
+          mtm_formula,
+          trading_period,
+          pricing_period_start,
+          loading_period_start,
+          pricing_type,
+          efp_designated_month
+        `).order('trading_period', {
+        ascending: true
+      });
+      if (physicalError) throw physicalError;
+
+      const {
+        data: paperTradeLegs,
+        error: paperError
+      } = await supabase.from('paper_trade_legs').select(`
+          id,
+          leg_reference,
+          buy_sell,
+          product,
+          quantity,
+          formula,
+          mtm_formula,
+          exposures,
+          period,
+          trading_period,
+          instrument
+        `).order('period', {
+        ascending: true
+      });
+      if (paperError) throw paperError;
+
+      return {
+        physicalTradeLegs: physicalTradeLegs || [],
+        paperTradeLegs: paperTradeLegs || []
+      };
+    }
+  });
+
+  // Process trade data into exposure data
+  const exposureData = useMemo(() => {
+    const exposuresByMonth: Record<string, Record<string, ExposureData>> = {};
+    const allProductsFound = new Set<string>();
+    periods.forEach(month => {
+      exposuresByMonth[month] = {};
+      ALLOWED_PRODUCTS.forEach(product => {
+        exposuresByMonth[month][product] = {
+          physical: 0,
+          pricing: 0,
+          paper: 0,
+          netExposure: 0
+        };
+      });
+    });
+
+    if (tradeData) {
+      const {
+        physicalTradeLegs,
+        paperTradeLegs
+      } = tradeData;
+
+      if (physicalTradeLegs && physicalTradeLegs.length > 0) {
+        physicalTradeLegs.forEach(leg => {
+          let physicalExposureMonth = '';
+          if (leg.loading_period_start) {
+            physicalExposureMonth = formatMonthCode(new Date(leg.loading_period_start));
+          } else if (leg.trading_period) {
+            physicalExposureMonth = leg.trading_period;
+          } else if (leg.pricing_period_start) {
+            physicalExposureMonth = formatMonthCode(new Date(leg.pricing_period_start));
+          }
+
+          let pricingExposureMonth = '';
+          if (leg.pricing_type === 'efp' && leg.efp_designated_month) {
+            pricingExposureMonth = leg.efp_designated_month;
+          } else if (leg.trading_period) {
+            pricingExposureMonth = leg.trading_period;
+          } else if (leg.pricing_period_start) {
+            pricingExposureMonth = formatMonthCode(new Date(leg.pricing_period_start));
+          }
+
+          if (physicalExposureMonth && periods.includes(physicalExposureMonth)) {
+            const canonicalProduct = mapProductToCanonical(leg.product || 'Unknown');
+            const quantityMultiplier = leg.buy_sell === 'buy' ? 1 : -1;
+            const quantity = (leg.quantity || 0) * quantityMultiplier;
+            allProductsFound.add(canonicalProduct);
+            if (!exposuresByMonth[physicalExposureMonth][canonicalProduct]) {
+              exposuresByMonth[physicalExposureMonth][canonicalProduct] = {
+                physical: 0,
+                pricing: 0,
+                paper: 0,
+                netExposure: 0
+              };
+            }
+            const mtmFormula = validateAndParsePricingFormula(leg.mtm_formula);
+            if (mtmFormula.tokens.length > 0) {
+              if (mtmFormula.exposures && mtmFormula.exposures.physical) {
+                Object.entries(mtmFormula.exposures.physical).forEach(([baseProduct, weight]) => {
+                  const canonicalBaseProduct = mapProductToCanonical(baseProduct);
+                  allProductsFound.add(canonicalBaseProduct);
+                  if (!exposuresByMonth[physicalExposureMonth][canonicalBaseProduct]) {
+                    exposuresByMonth[physicalExposureMonth][canonicalBaseProduct] = {
+                      physical: 0,
+                      pricing: 0,
+                      paper: 0,
+                      netExposure: 0
+                    };
+                  }
+                  const actualExposure = typeof weight === 'number' ? weight : 0;
+                  exposuresByMonth[physicalExposureMonth][canonicalBaseProduct].physical += actualExposure;
+                });
+              } else {
+                exposuresByMonth[physicalExposureMonth][canonicalProduct].physical += quantity;
+              }
+            } else {
+              exposuresByMonth[physicalExposureMonth][canonicalProduct].physical += quantity;
+            }
+          }
+
+          const pricingFormula = validateAndParsePricingFormula(leg.pricing_formula);
+
+          if (pricingFormula.monthlyDistribution) {
+            Object.entries(pricingFormula.monthlyDistribution).forEach(([instrument, monthlyValues]) => {
+              const canonicalInstrument = mapProductToCanonical(instrument);
+              allProductsFound.add(canonicalInstrument);
+              Object.entries(monthlyValues).forEach(([monthCode, value]) => {
+                if (periods.includes(monthCode) && value !== 0) {
+                  if (!exposuresByMonth[monthCode][canonicalInstrument]) {
+                    exposuresByMonth[monthCode][canonicalInstrument] = {
+                      physical: 0,
+                      pricing: 0,
+                      paper: 0,
+                      netExposure: 0
+                    };
+                  }
+                  exposuresByMonth[monthCode][canonicalInstrument].pricing += value;
+                }
+              });
+            });
+          } else if (pricingExposureMonth && periods.includes(pricingExposureMonth) && pricingFormula.exposures && pricingFormula.exposures.pricing) {
+            Object.entries(pricingFormula.exposures.pricing).forEach(([instrument, value]) => {
+              const canonicalInstrument = mapProductToCanonical(instrument);
+              allProductsFound.add(canonicalInstrument);
+              if (!exposuresByMonth[pricingExposureMonth][canonicalInstrument]) {
+                exposuresByMonth[pricingExposureMonth][canonicalInstrument] = {
+                  physical: 0,
+                  pricing: 0,
+                  paper: 0,
+                  netExposure: 0
+                };
+              }
+              exposuresByMonth[pricingExposureMonth][canonicalInstrument].pricing += Number(value) || 0;
+            });
+          }
+        });
+      }
+
+      if (paperTradeLegs && paperTradeLegs.length > 0) {
+        paperTradeLegs.forEach(leg => {
+          const month = leg.period || leg.trading_period || '';
+          if (!month || !periods.includes(month)) {
+            return;
+          }
+          if (leg.instrument) {
+            const {
+              baseProduct,
+              oppositeProduct,
+              relationshipType
+            } = parsePaperInstrument(leg.instrument);
+            if (baseProduct) {
+              allProductsFound.add(baseProduct);
+              if (!exposuresByMonth[month][baseProduct]) {
+                exposuresByMonth[month][baseProduct] = {
+                  physical: 0,
+                  pricing: 0,
+                  paper: 0,
+                  netExposure: 0
+                };
+              }
+              const buySellMultiplier = leg.buy_sell === 'buy' ? 1 : -1;
+              const quantity = (leg.quantity || 0) * buySellMultiplier;
+              exposuresByMonth[month][baseProduct].paper += quantity;
+              exposuresByMonth[month][baseProduct].pricing += quantity;
+              if ((relationshipType === 'DIFF' || relationshipType === 'SPREAD') && oppositeProduct) {
+                allProductsFound.add(oppositeProduct);
+                if (!exposuresByMonth[month][oppositeProduct]) {
+                  exposuresByMonth[month][oppositeProduct] = {
+                    physical: 0,
+                    pricing: 0,
+                    paper: 0,
+                    netExposure: 0
+                  };
+                }
+                exposuresByMonth[month][oppositeProduct].paper += -quantity;
+                exposuresByMonth[month][oppositeProduct].pricing += -quantity;
+              }
+            } else if (leg.exposures && typeof leg.exposures === 'object') {
+              const exposuresData = leg.exposures as Record<string, any>;
+              if (exposuresData.physical && typeof exposuresData.physical === 'object') {
+                Object.entries(exposuresData.physical).forEach(([prodName, value]) => {
+                  const canonicalProduct = mapProductToCanonical(prodName);
+                  allProductsFound.add(canonicalProduct);
+                  if (!exposuresByMonth[month][canonicalProduct]) {
+                    exposuresByMonth[month][canonicalProduct] = {
+                      physical: 0,
+                      pricing: 0,
+                      netExposure: 0,
+                      paper: 0
+                    };
+                  }
+                  exposuresByMonth[month][canonicalProduct].paper += Number(value) || 0;
+                  if (!exposuresData.pricing || typeof exposuresData.pricing !== 'object' || !exposuresData.pricing[prodName]) {
+                    exposuresByMonth[month][canonicalProduct].pricing += Number(value) || 0;
+                  }
+                });
+              }
+              if (exposuresData.pricing && typeof exposuresData.pricing === 'object') {
+                Object.entries(exposuresData.pricing).forEach(([instrument, value]) => {
+                  const canonicalInstrument = mapProductToCanonical(instrument);
+                  allProductsFound.add(canonicalInstrument);
+                  if (!exposuresByMonth[month][canonicalInstrument]) {
+                    exposuresByMonth[month][canonicalInstrument] = {
+                      physical: 0,
+                      pricing: 0,
+                      paper: 0,
+                      netExposure: 0
+                    };
+                  }
+                  exposuresByMonth[month][canonicalInstrument].pricing += Number(value) || 0;
+                });
+              }
+            } else if (leg.mtm_formula && typeof leg.mtm_formula === 'object') {
+              const mtmFormula = leg.mtm_formula as Record<string, any>;
+              if (mtmFormula.exposures && typeof mtmFormula.exposures === 'object') {
+                const mtmExposures = mtmFormula.exposures as Record<string, any>;
+                if (mtmExposures.physical && typeof mtmExposures.physical === 'object') {
+                  Object.entries(mtmExposures.physical).forEach(([prodName, value]) => {
+                    const canonicalProduct = mapProductToCanonical(prodName);
+                    allProductsFound.add(canonicalProduct);
+                    if (!exposuresByMonth[month][canonicalProduct]) {
+                      exposuresByMonth[month][canonicalProduct] = {
+                        physical: 0,
+                        pricing: 0,
+                        paper: 0,
+                        netExposure: 0
+                      };
+                    }
+                    const paperExposure = Number(value) || 0;
+                    exposuresByMonth[month][canonicalProduct].paper += paperExposure;
+                    if (!mtmExposures.pricing || !(prodName in (mtmExposures.pricing || {}))) {
+                      exposuresByMonth[month][canonicalProduct].pricing += paperExposure;
+                    }
+                  });
+                }
+                if (mtmExposures.pricing && typeof mtmExposures.pricing === 'object') {
+                  Object.entries(mtmExposures.pricing).forEach(([prodName, value]) => {
+                    const canonicalProduct = mapProductToCanonical(prodName);
+                    allProductsFound.add(canonicalProduct);
+                    if (!exposuresByMonth[month][canonicalProduct]) {
+                      exposuresByMonth[month][canonicalProduct] = {
+                        physical: 0,
+                        pricing: 0,
+                        paper: 0,
+                        netExposure: 0
+                      };
+                    }
+                    exposuresByMonth[month][canonicalProduct].pricing += Number(value) || 0;
+                  });
+                }
+              }
+            } else {
+              const canonicalProduct = mapProductToCanonical(leg.product || 'Unknown');
+              if (leg.mtm_formula && typeof leg.mtm_formula === 'object') {
+                const mtmFormula = validateAndParsePricingFormula(leg.mtm_formula);
+                if (mtmFormula.exposures && mtmFormula.exposures.physical && Object.keys(mtmFormula.exposures.physical).length > 0) {
+                  const buySellMultiplier = leg.buy_sell === 'buy' ? 1 : -1;
+                  Object.entries(mtmFormula.exposures.physical).forEach(([pBaseProduct, weight]) => {
+                    const canonicalBaseProduct = mapProductToCanonical(pBaseProduct);
+                    allProductsFound.add(canonicalBaseProduct);
+                    if (!exposuresByMonth[month][canonicalBaseProduct]) {
+                      exposuresByMonth[month][canonicalBaseProduct] = {
+                        physical: 0,
+                        pricing: 0,
+                        paper: 0,
+                        netExposure: 0
+                      };
+                    }
+                    const actualExposure = typeof weight === 'number' ? weight * buySellMultiplier : 0;
+                    exposuresByMonth[month][canonicalBaseProduct].paper += actualExposure;
+                    if (!mtmFormula.exposures.pricing || !(pBaseProduct in (mtmFormula.exposures.pricing || {}))) {
+                      exposuresByMonth[month][canonicalBaseProduct].pricing += actualExposure;
+                    }
+                  });
+                  if (mtmFormula.exposures.pricing) {
+                    Object.entries(mtmFormula.exposures.pricing).forEach(([pBaseProduct, weight]) => {
+                      const canonicalBaseProduct = mapProductToCanonical(pBaseProduct);
+                      allProductsFound.add(canonicalBaseProduct);
+                      if (!exposuresByMonth[month][canonicalBaseProduct]) {
+                        exposuresByMonth[month][canonicalBaseProduct] = {
+                          physical: 0,
+                          pricing: 0,
+                          paper: 0,
+                          netExposure: 0
+                        };
+                      }
+                      const actualExposure = typeof weight === 'number' ? weight * buySellMultiplier : 0;
+                      exposuresByMonth[month][canonicalBaseProduct].pricing += actualExposure;
+                    });
+                  }
+                } else {
+                  if (!exposuresByMonth[month][canonicalProduct]) {
+                    exposuresByMonth[month][canonicalProduct] = {
+                      physical: 0,
+                      pricing: 0,
+                      paper: 0,
+                      netExposure: 0
+                    };
+                  }
+                  const buySellMultiplier = leg.buy_sell === 'buy' ? 1 : -1;
+                  const paperExposure = (leg.quantity || 0) * buySellMultiplier;
+                  exposuresByMonth[month][canonicalProduct].paper += paperExposure;
+                  exposuresByMonth[month][canonicalProduct].pricing += paperExposure;
+                }
+              } else {
+                if (!exposuresByMonth[month][canonicalProduct]) {
+                  exposuresByMonth[month][canonicalProduct] = {
+                    physical: 0,
+                    pricing: 0,
+                    paper: 0,
+                    netExposure: 0
+                  };
+                }
+                const buySellMultiplier = leg.buy_sell === 'buy' ? 1 : -1;
+                const paperExposure = (leg.quantity || 0) * buySellMultiplier;
+                exposuresByMonth[month][canonicalProduct].paper += paperExposure;
+                exposuresByMonth[month][canonicalProduct].pricing += paperExposure;
+              }
+            }
+          } else if (leg.exposures && typeof leg.exposures === 'object') {
+            const exposuresData = leg.exposures as Record<string, any>;
+            if (exposuresData.physical && typeof exposuresData.physical === 'object') {
+              Object.entries(exposuresData.physical).forEach(([prodName, value]) => {
+                const canonicalProduct = mapProductToCanonical(prodName);
+                allProductsFound.add(canonicalProduct);
+                if (!exposuresByMonth[month][canonicalProduct]) {
+                  exposuresByMonth[month][canonicalProduct] = {
+                    physical: 0,
+                    pricing: 0,
+                    netExposure: 0,
+                    paper: 0
+                  };
+                }
+                exposuresByMonth[month][canonicalProduct].paper += Number(value) || 0;
+                if (!exposuresData.pricing || typeof exposuresData.pricing !== 'object' || !exposuresData.pricing[prodName]) {
+                  exposuresByMonth[month][canonicalProduct].pricing += Number(value) || 0;
+                }
+              });
+            }
+            if (exposuresData.pricing && typeof exposuresData.pricing === 'object') {
+              Object.entries(exposuresData.pricing).forEach(([instrument, value]) => {
+                const canonicalInstrument = mapProductToCanonical(instrument);
+                allProductsFound.add(canonicalInstrument);
+                if (!exposuresByMonth[month][canonicalInstrument]) {
+                  exposuresByMonth[month][canonicalInstrument] = {
+                    physical: 0,
+                    pricing: 0,
+                    paper: 0,
+                    netExposure: 0
+                  };
+                }
+                exposuresByMonth[month][canonicalInstrument].pricing += Number(value) || 0;
+              });
+            }
+          } else if (leg.mtm_formula && typeof leg.mtm_formula === 'object') {
+            const mtmFormula = leg.mtm_formula as Record<string, any>;
+            if (mtmFormula.exposures && typeof mtmFormula.exposures === 'object') {
+              const mtmExposures = mtmFormula.exposures as Record<string, any>;
+              if (mtmExposures.physical && typeof mtmExposures.physical === 'object') {
+                Object.entries(mtmExposures.physical).forEach(([prodName, value]) => {
+                  const canonicalProduct = mapProductToCanonical(prodName);
+                  allProductsFound.add(canonicalProduct);
+                  if (!exposuresByMonth[month][canonicalProduct]) {
+                    exposuresByMonth[month][canonicalProduct] = {
+                      physical: 0,
+                      pricing: 0,
+                      paper: 0,
+                      netExposure: 0
+                    };
+                  }
+                  const paperExposure = Number(value) || 0;
+                  exposuresByMonth[month][canonicalProduct].paper += paperExposure;
+                  if (!mtmExposures.pricing || !(prodName in (mtmExposures.pricing || {}))) {
+                    exposuresByMonth[month][canonicalProduct].pricing += paperExposure;
+                  }
+                });
+              }
+              if (mtmExposures.pricing && typeof mtmExposures.pricing === 'object') {
+                Object.entries(mtmExposures.pricing).forEach(([prodName, value]) => {
+                  const canonicalProduct = mapProductToCanonical(prodName);
+                  allProductsFound.add(canonicalProduct);
+                  if (!exposuresByMonth[month][canonicalProduct]) {
+                    exposuresByMonth[month][canonicalProduct] = {
+                      physical: 0,
+                      pricing: 0,
+                      paper: 0,
+                      netExposure: 0
+                    };
+                  }
+                  exposuresByMonth[month][canonicalProduct].pricing += Number(value) || 0;
+                });
+              }
+            }
+          } else {
+            const canonicalProduct = mapProductToCanonical(leg.product || 'Unknown');
+            if (leg.mtm_formula && typeof leg.mtm_formula === 'object') {
+              const mtmFormula = validateAndParsePricingFormula(leg.mtm_formula);
+              if (mtmFormula.exposures && mtmFormula.exposures.physical && Object.keys(mtmFormula.exposures.physical).length > 0) {
+                const buySellMultiplier = leg.buy_sell === 'buy' ? 1 : -1;
+                Object.entries(mtmFormula.exposures.physical).forEach(([pBaseProduct, weight]) => {
+                  const canonicalBaseProduct = mapProductToCanonical(pBaseProduct);
+                  allProductsFound.add(canonicalBaseProduct);
+                  if (!exposuresByMonth[month][canonicalBaseProduct]) {
+                    exposuresByMonth[month][canonicalBaseProduct] = {
+                      physical: 0,
+                      pricing: 0,
+                      paper: 0,
+                      netExposure: 0
+                    };
+                  }
+                  const actualExposure = typeof weight === 'number' ? weight * buySellMultiplier : 0;
+                  exposuresByMonth[month][canonicalBaseProduct].paper += actualExposure;
+                  if (!mtmFormula.exposures.pricing || !(pBaseProduct in (mtmFormula.exposures.pricing || {}))) {
+                    exposuresByMonth[month][canonicalBaseProduct].pricing += actualExposure;
+                  }
+                });
+                if (mtmFormula.exposures.pricing) {
+                  Object.entries(mtmFormula.exposures.pricing).forEach(([pBaseProduct, weight]) => {
+                    const canonicalBaseProduct = mapProductToCanonical(pBaseProduct);
+                    allProductsFound.add(canonicalBaseProduct);
+                    if (!exposuresByMonth[month][canonicalBaseProduct]) {
+                      exposuresByMonth[month][canonicalBaseProduct] = {
+                        physical: 0,
+                        pricing: 0,
+                        paper: 0,
+                        netExposure: 0
+                      };
+                    }
+                    const actualExposure = typeof weight === 'number' ? weight * buySellMultiplier : 0;
+                    exposuresByMonth[month][canonicalBaseProduct].pricing += actualExposure;
+                  });
+                }
+              } else {
+                if (!exposuresByMonth[month][canonicalProduct]) {
+                  exposuresByMonth[month][canonicalProduct] = {
+                    physical: 0,
+                    pricing: 0,
+                    paper: 0,
+                    netExposure: 0
+                  };
+                }
+                const buySellMultiplier = leg.buy_sell === 'buy' ? 1 : -1;
+                const paperExposure = (leg.quantity || 0) * buySellMultiplier;
+                exposuresByMonth[month][canonicalProduct].paper += paperExposure;
+                exposuresByMonth[month][canonicalProduct].pricing += paperExposure;
+              }
+            } else {
+              if (!exposuresByMonth[month][canonicalProduct]) {
+                exposuresByMonth[month][canonicalProduct] = {
+                  physical: 0,
+                  pricing: 0,
+                  paper: 0,
+                  netExposure: 0
+                };
+              }
+              const buySellMultiplier = leg.buy_sell === 'buy' ? 1 : -1;
+              const paperExposure = (leg.quantity || 0) * buySellMultiplier;
+              exposuresByMonth[month][canonicalProduct].paper += paperExposure;
+              exposuresByMonth[month][canonicalProduct].pricing += paperExposure;
+            }
+          }
+        });
+      }
+    }
+
+    // Calculate net exposure
+    periods.forEach(month => {
+      Object.entries(exposuresByMonth[month]).forEach(([product, exposure]) => {
+        exposure.netExposure = calculateNetExposure(exposure.physical, exposure.pricing);
+      });
+    });
+
+    // Format the data into the expected structure
+    const monthlyExposures: MonthlyExposure[] = periods.map(month => {
+      const monthData = exposuresByMonth[month];
+      const productsData: Record<string, ExposureData> = {};
+      const totals: ExposureData = {
+        physical: 0,
+        pricing: 0,
+        paper: 0,
+        netExposure: 0
+      };
+
+      Object.entries(monthData).forEach(([product, exposure]) => {
+        if (ALLOWED_PRODUCTS.includes(product)) {
+          productsData[product] = exposure;
+          totals.physical += exposure.physical;
+          totals.pricing += exposure.pricing;
+          totals.paper += exposure.paper;
+        }
+      });
+
+      totals.netExposure = calculateNetExposure(totals.physical, totals.pricing);
+      return {
+        month,
+        products: productsData,
+        totals
+      };
+    });
+
+    return monthlyExposures;
+  }, [tradeData, periods, ALLOWED_PRODUCTS]);
+
+  // Get all products for selection
+  const allProducts = useMemo(() => {
+    return [...ALLOWED_PRODUCTS].sort();
+  }, [ALLOWED_PRODUCTS]);
+
+  // Update selected products when allProducts changes
+  React.useEffect(() => {
+    if (allProducts.length > 0) {
+      setSelectedProducts([...allProducts]);
+    }
+  }, [allProducts]);
+
+  const toggleCategory = (category: string) => {
+    setVisibleCategories(prev => {
+      if (prev.includes(category)) {
+        return prev.filter(cat => cat !== category);
+      } else {
+        const newCategories = [...prev, category];
+        return ['Physical', 'Pricing', 'Paper', 'Exposure'].filter(cat => newCategories.includes(cat));
+      }
+    });
+  };
+
+  return {
+    periods,
+    visibleCategories,
+    setVisibleCategories,
+    toggleCategory,
+    selectedProducts,
+    setSelectedProducts,
+    allProducts,
+    exposureData,
+    isLoading,
+    error,
+    refetch,
+    instrumentsLoading,
+    BIODIESEL_PRODUCTS,
+    PRICING_INSTRUMENT_PRODUCTS,
+    ALLOWED_PRODUCTS
+  };
+};
